Improve documentation
Edit README
Add ESDoc comments for parser classes
Add ESDoc comments for error classes
Add ESDoc comments for chart types
