+++ .esdoc.json
@@ -3,7 +3,30 @@
   "destination": "./dist/docs",
   "plugins": [
     {
-      "name": "esdoc-standard-plugin"
+      "name": "esdoc-standard-plugin",
+      "option": {
+        "lint": {"enable": true},
+        "coverage": {"enable": true},
+        "accessor": {"access": ["public", "protected", "private"], "autoPrivate": true},
+        "undocumentIdentifier": {"enable": true},
+        "unexportedIdentifier": {"enable": false},
+        "typeInference": {"enable": true},
+        "brand": {
+          "logo": "./Resources/Original/analytics_512.png",
+          "title": "GraphJS",
+          "description": "A simple grammar to convert raw data into visual graphs",
+          "repository": "https://github.com/OshanIvantha/GraphJS",
+          "site": "http://ivantha.me/GraphJS",
+          "author": "https://twitter.com/OshanIvantha",
+          "image": "./Resources/Original/analytics_512.png"
+        },
+        "test": {
+          "source": "./test/",
+          "interfaces": ["describe", "it", "context", "suite", "test"],
+          "includes": ["(spec|Spec|test|Test)\\.js$"],
+          "excludes": ["\\.config\\.js$"]
+        }
+      }
     }
   ]
 }
