--- src/chart/combination-chart.js
@@ -1,8 +1,0 @@
-import Chart from './chart'
-
-export default class CombinationChart extends Chart {
-
-	constructor(data) {
-		super(data)
-	}
-}
