--- src/chart/gauge-chart.js
@@ -1,8 +1,0 @@
-import Chart from './chart'
-
-export default class GaugeChart extends Chart {
-
-	constructor(data) {
-		super(data)
-	}
-}
