--- src/chart/line-chart-with-regions.js
@@ -1,8 +1,0 @@
-import Chart from './chart'
-
-export default class LineChartWithRegions extends Chart {
-
-	constructor(data) {
-		super(data)
-	}
-}
