--- src/chart/line-chart.js
@@ -1,9 +1,0 @@
-import Chart from './chart'
-
-export default class LineChart extends Chart {
-
-	constructor(data) {
-		super(data)
-	}
-
-}
