+++ src/charts/types/combination-chart.js
@@ -1,0 +1,15 @@
+import Chart from '../chart'
+
+/**
+ * Represents a Combination Chart.
+ */
+export default class CombinationChart extends Chart {
+
+	/**
+	 * Creates an instance of the CombinationChart class
+	 * @param {string} data - JSON string containing the data columns.
+	 */
+	constructor(data) {
+		super(data)
+	}
+}
