+++ src/charts/types/line-chart-with-regions.js
@@ -1,0 +1,15 @@
+import Chart from '../chart'
+
+/**
+ * Represents a Line Chart with Regions.
+ */
+export default class LineChartWithRegions extends Chart {
+
+	/**
+	 * Creates an instance of the LineChartWithRegions class
+	 * @param {string} data - JSON string containing the data columns.
+	 */
+	constructor(data) {
+		super(data)
+	}
+}
