+++ src/charts/types/line-chart.js
@@ -1,0 +1,16 @@
+import Chart from '../chart'
+
+/**
+ * Represents a Line Chart.
+ */
+export default class LineChart extends Chart {
+
+	/**
+	 * Creates an instance of the LineChart class
+	 * @param {string} data - JSON string containing the data columns.
+	 */
+	constructor(data) {
+		super(data)
+	}
+
+}
