--- src/parser/csv-parser.js
@@ -1,8 +1,0 @@
-/**
- * Parse CSV files into the desired format of GraphJS library
- */
-export default class CsvParser {
-
-	constructor() {
-	}
-}
