--- src/parser/json-parser.js
@@ -1,8 +1,0 @@
-/**
- * Parse JSON files into the desired format of GraphJS library
- */
-export default class JsonParser {
-
-	constructor() {
-	}
-}
