--- src/parser/tsv-parser.js
@@ -1,8 +1,0 @@
-/**
- * Parse TSV files into the desired format of GraphJS library
- */
-export default class TsvParser {
-
-	constructor() {
-	}
-}
