--- src/parser/txt-parser.js
@@ -1,8 +1,0 @@
-/**
- * Parse TXT files into the desired format of GraphJS library
- */
-export default class TxtParser {
-
-	constructor() {
-	}
-}
