+++ src/parsers/csv-parser.js
@@ -1,0 +1,12 @@
+/**
+ * Parse CSV files into the desired format of GraphJS library.
+ */
+// TODO : To be implemented
+export default class CsvParser {
+
+	/**
+	 * Creates an instance of the CsvParser class
+	 */
+	constructor() {
+	}
+}
