+++ src/parsers/json-parser.js
@@ -1,0 +1,12 @@
+/**
+ * Parse JSON files into the desired format of GraphJS library.
+ */
+// TODO : To be implemented
+export default class JsonParser {
+
+	/**
+	 * Creates an instance of the JsonParser class
+	 */
+	constructor() {
+	}
+}
