+++ src/parsers/tsv-parser.js
@@ -1,0 +1,12 @@
+/**
+ * Parse TSV files into the desired format of GraphJS library.
+ */
+// TODO : To be implemented
+export default class TsvParser {
+
+	/**
+	 * Creates an instance of the TsvParser class
+	 */
+	constructor() {
+	}
+}
