+++ src/parsers/txt-parser.js
@@ -1,0 +1,12 @@
+/**
+ * Parse TXT files into the desired format of GraphJS library.
+ */
+// TODO : To be implemented
+export default class TxtParser {
+
+	/**
+	 * Creates an instance of the TxtParser class
+	 */
+	constructor() {
+	}
+}
